+++ src/Pages/Boards/index.jsx
@@ -1,0 +1,1 @@
+// Boards list
