+++ src/Pages/Users/_id.jsx
@@ -1,0 +1,1 @@
+// User Details
